chore: cleanup all old models in pkg/models except for old Canvas
