+++ docker-compose.dev.yml
@@ -30,8 +30,6 @@ services:
       PUBLIC_API_BASE_PATH: /api/v1
       START_WEB_SERVER: "yes"
       START_EVENT_DISTRIBUTER: "yes"
-      START_HARD_DELETION_WORKER: "yes"
-      START_EVENT_DELETION_WORKER: "yes"
       START_WORKFLOW_EVENT_ROUTER: "yes"
       START_WORKFLOW_NODE_EXECUTOR: "yes"
       START_BLUEPRINT_NODE_EXECUTOR: "yes"
