+++ pkg/grpc/actions/common.go
@@ -3,17 +3,13 @@ package actions
 import (
 	"context"
 	"encoding/json"
-	"errors"
 	"fmt"
-	"sort"
-	"time"
 
 	uuid "github.com/google/uuid"
 	"github.com/superplanehq/superplane/pkg/components"
 	"github.com/superplanehq/superplane/pkg/integrations"
 	"github.com/superplanehq/superplane/pkg/models"
 	pbAuth "github.com/superplanehq/superplane/pkg/protos/authorization"
-	pb "github.com/superplanehq/superplane/pkg/protos/canvases"
 	componentpb "github.com/superplanehq/superplane/pkg/protos/components"
 	integrationpb "github.com/superplanehq/superplane/pkg/protos/integrations"
 	"github.com/superplanehq/superplane/pkg/registry"
@@ -35,272 +31,6 @@ func ValidateUUIDsArray(ids []string) error {
 	}
 
 	return nil
-}
-
-func ExecutionResultToProto(result string) pb.Execution_Result {
-	switch result {
-	case models.ResultFailed:
-		return pb.Execution_RESULT_FAILED
-	case models.ResultPassed:
-		return pb.Execution_RESULT_PASSED
-	case models.ResultCancelled:
-		return pb.Execution_RESULT_CANCELLED
-	default:
-		return pb.Execution_RESULT_UNKNOWN
-	}
-}
-
-func ExecutionResultReasonToProto(reason string) pb.Execution_ResultReason {
-	switch reason {
-	case models.ResultReasonError:
-		return pb.Execution_RESULT_REASON_ERROR
-	case models.ResultReasonMissingOutputs:
-		return pb.Execution_RESULT_REASON_MISSING_OUTPUTS
-	case models.ResultReasonTimeout:
-		return pb.Execution_RESULT_REASON_TIMEOUT
-	case models.ResultReasonUser:
-		return pb.Execution_RESULT_REASON_USER
-	default:
-		return pb.Execution_RESULT_REASON_OK
-	}
-}
-
-func FindConnectionSourceID(canvasID string, connection *pb.Connection) (*uuid.UUID, error) {
-	switch connection.Type {
-	case pb.Connection_TYPE_STAGE:
-		stage, err := models.FindStageByName(canvasID, connection.Name)
-		if err != nil {
-			return nil, fmt.Errorf("stage %s not found", connection.Name)
-		}
-
-		return &stage.ID, nil
-
-	case pb.Connection_TYPE_EVENT_SOURCE:
-		eventSource, err := models.FindExternalEventSourceByName(canvasID, connection.Name)
-		if err != nil {
-			return nil, fmt.Errorf("event source %s not found", connection.Name)
-		}
-
-		return &eventSource.ID, nil
-
-	case pb.Connection_TYPE_CONNECTION_GROUP:
-		connectionGroup, err := models.FindConnectionGroupByName(canvasID, connection.Name)
-		if err != nil {
-			return nil, fmt.Errorf("connection group %s not found", connection.Name)
-		}
-
-		return &connectionGroup.ID, nil
-
-	default:
-		return nil, errors.New("invalid type")
-	}
-}
-
-func ValidateConnections(canvasID string, connections []*pb.Connection) ([]models.Connection, error) {
-	cs := []models.Connection{}
-
-	if len(connections) == 0 {
-		return nil, fmt.Errorf("connections must not be empty")
-	}
-
-	for _, connection := range connections {
-		sourceID, err := FindConnectionSourceID(canvasID, connection)
-		if err != nil {
-			return nil, fmt.Errorf("invalid connection: %v", err)
-		}
-
-		filters, err := ValidateFilters(connection.Filters)
-		if err != nil {
-			return nil, err
-		}
-
-		cs = append(cs, models.Connection{
-			SourceID:       *sourceID,
-			SourceName:     connection.Name,
-			SourceType:     ProtoToConnectionType(connection.Type),
-			FilterOperator: ProtoToFilterOperator(connection.FilterOperator),
-			Filters:        filters,
-		})
-	}
-
-	return cs, nil
-}
-
-func ValidateFilters(in []*pb.Filter) ([]models.Filter, error) {
-	filters := []models.Filter{}
-	for i, f := range in {
-		filter, err := validateFilter(f)
-		if err != nil {
-			return nil, fmt.Errorf("invalid filter [%d]: %v", i, err)
-		}
-
-		filters = append(filters, *filter)
-	}
-
-	return filters, nil
-}
-
-func validateFilter(filter *pb.Filter) (*models.Filter, error) {
-	switch filter.Type {
-	case pb.FilterType_FILTER_TYPE_DATA:
-		return validateDataFilter(filter.Data)
-	case pb.FilterType_FILTER_TYPE_HEADER:
-		return validateHeaderFilter(filter.Header)
-	default:
-		return nil, fmt.Errorf("invalid filter type: %s", filter.Type)
-	}
-}
-
-func validateDataFilter(filter *pb.DataFilter) (*models.Filter, error) {
-	if filter == nil {
-		return nil, fmt.Errorf("no filter provided")
-	}
-
-	if filter.Expression == "" {
-		return nil, fmt.Errorf("expression is empty")
-	}
-
-	return &models.Filter{
-		Type: models.FilterTypeData,
-		Data: &models.DataFilter{
-			Expression: filter.Expression,
-		},
-	}, nil
-}
-
-func validateHeaderFilter(filter *pb.HeaderFilter) (*models.Filter, error) {
-	if filter == nil {
-		return nil, fmt.Errorf("no filter provided")
-	}
-
-	if filter.Expression == "" {
-		return nil, fmt.Errorf("expression is empty")
-	}
-
-	return &models.Filter{
-		Type: models.FilterTypeHeader,
-		Header: &models.HeaderFilter{
-			Expression: filter.Expression,
-		},
-	}, nil
-}
-
-func ProtoToConnectionType(t pb.Connection_Type) string {
-	switch t {
-	case pb.Connection_TYPE_STAGE:
-		return models.SourceTypeStage
-	case pb.Connection_TYPE_EVENT_SOURCE:
-		return models.SourceTypeEventSource
-	case pb.Connection_TYPE_CONNECTION_GROUP:
-		return models.SourceTypeConnectionGroup
-	default:
-		return ""
-	}
-}
-
-func ProtoToFilterOperator(in pb.FilterOperator) string {
-	switch in {
-	case pb.FilterOperator_FILTER_OPERATOR_OR:
-		return models.FilterOperatorOr
-	default:
-		return models.FilterOperatorAnd
-	}
-}
-
-func FilterOperatorToProto(in string) pb.FilterOperator {
-	switch in {
-	case models.FilterOperatorOr:
-		return pb.FilterOperator_FILTER_OPERATOR_OR
-	default:
-		return pb.FilterOperator_FILTER_OPERATOR_AND
-	}
-}
-
-func SerializeConnections(in []models.Connection) ([]*pb.Connection, error) {
-	connections := []*pb.Connection{}
-
-	for _, c := range in {
-		filters, err := SerializeFilters(c.Filters)
-		if err != nil {
-			return nil, fmt.Errorf("invalid filters: %v", err)
-		}
-
-		connections = append(connections, &pb.Connection{
-			Type:           ConnectionTypeToProto(c.SourceType),
-			Name:           c.SourceName,
-			FilterOperator: FilterOperatorToProto(c.FilterOperator),
-			Filters:        filters,
-		})
-	}
-
-	//
-	// Sort them by name so we have some predictability here.
-	//
-	sort.SliceStable(connections, func(i, j int) bool {
-		return connections[i].Name < connections[j].Name
-	})
-
-	return connections, nil
-}
-
-func SerializeFilters(in []models.Filter) ([]*pb.Filter, error) {
-	filters := []*pb.Filter{}
-
-	for _, f := range in {
-		filter, err := SerializeFilter(f)
-		if err != nil {
-			return nil, fmt.Errorf("invalid filter: %v", err)
-		}
-
-		filters = append(filters, filter)
-	}
-
-	return filters, nil
-}
-
-func SerializeFilter(in models.Filter) (*pb.Filter, error) {
-	switch in.Type {
-	case models.FilterTypeData:
-		return &pb.Filter{
-			Type: pb.FilterType_FILTER_TYPE_DATA,
-			Data: &pb.DataFilter{
-				Expression: in.Data.Expression,
-			},
-		}, nil
-	case models.FilterTypeHeader:
-		return &pb.Filter{
-			Type: pb.FilterType_FILTER_TYPE_HEADER,
-			Header: &pb.HeaderFilter{
-				Expression: in.Header.Expression,
-			},
-		}, nil
-	default:
-		return nil, fmt.Errorf("invalid filter type: %s", in.Type)
-	}
-}
-
-func ConnectionTypeToProto(t string) pb.Connection_Type {
-	switch t {
-	case models.SourceTypeStage:
-		return pb.Connection_TYPE_STAGE
-	case models.SourceTypeEventSource:
-		return pb.Connection_TYPE_EVENT_SOURCE
-	case models.SourceTypeConnectionGroup:
-		return pb.Connection_TYPE_CONNECTION_GROUP
-	default:
-		return pb.Connection_TYPE_UNKNOWN
-	}
-}
-
-func RejectionReasonToProto(reason string) pb.EventRejection_RejectionReason {
-	switch reason {
-	case models.EventRejectionReasonFiltered:
-		return pb.EventRejection_REJECTION_REASON_FILTERED
-	case models.EventRejectionReasonError:
-		return pb.EventRejection_REJECTION_REASON_ERROR
-	default:
-		return pb.EventRejection_REJECTION_REASON_UNKNOWN
-	}
 }
 
 func ProtoToDomainType(domainType pbAuth.DomainType) (string, error) {
@@ -414,45 +144,6 @@ func GetDomainForSecret(domainTypeForResource string, domainIdForResource *uuid.
 	}
 
 	return models.DomainTypeOrganization, &canvas.OrganizationID, nil
-}
-
-func ValidateTime(timeStr string) error {
-	if timeStr == "" {
-		return status.Error(codes.InvalidArgument, "time is required")
-	}
-
-	_, err := time.Parse("15:04", timeStr)
-	if err != nil {
-		return status.Error(codes.InvalidArgument, "time must be in HH:MM format (24-hour)")
-	}
-
-	return nil
-}
-
-func StageEventStateToProto(state string) pb.StageEvent_State {
-	switch state {
-	case models.StageEventStatePending:
-		return pb.StageEvent_STATE_PENDING
-	case models.StageEventStateWaiting:
-		return pb.StageEvent_STATE_WAITING
-	case models.StageEventStateProcessed:
-		return pb.StageEvent_STATE_PROCESSED
-	case models.StageEventStateDiscarded:
-		return pb.StageEvent_STATE_DISCARDED
-	default:
-		return pb.StageEvent_STATE_UNKNOWN
-	}
-}
-
-func StageEventStateReasonToProto(stateReason string) pb.StageEvent_StateReason {
-	switch stateReason {
-	case models.StageEventStateReasonApproval:
-		return pb.StageEvent_STATE_REASON_APPROVAL
-	case models.StageEventStateReasonTimeWindow:
-		return pb.StageEvent_STATE_REASON_TIME_WINDOW
-	default:
-		return pb.StageEvent_STATE_REASON_UNKNOWN
-	}
 }
 
 func numberTypeOptionsToProto(opts *components.NumberTypeOptions) *componentpb.NumberTypeOptions {
