+++ pkg/public/server_test.go
@@ -9,18 +9,13 @@ import (
 	"os"
 	"path/filepath"
 	"testing"
-	"time"
 
-	"github.com/google/uuid"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 	"github.com/superplanehq/superplane/pkg/authorization"
-	"github.com/superplanehq/superplane/pkg/builders"
 	"github.com/superplanehq/superplane/pkg/crypto"
 	"github.com/superplanehq/superplane/pkg/jwt"
-	"github.com/superplanehq/superplane/pkg/models"
 	"github.com/superplanehq/superplane/pkg/registry"
-	"github.com/superplanehq/superplane/test/support"
 )
 
 func Test__HealthCheckEndpoint(t *testing.T) {
@@ -39,214 +34,6 @@ func Test__HealthCheckEndpoint(t *testing.T) {
 	})
 
 	require.Equal(t, 200, response.Code)
-}
-
-func Test__HandleExecutionOutputs(t *testing.T) {
-	r := support.SetupWithOptions(t, support.SetupOptions{
-		Source:      true,
-		Integration: true,
-	})
-
-	authService, err := authorization.NewAuthService()
-	require.NoError(t, err)
-	authService.EnableCache(false)
-
-	executorType, executorSpec, resource := support.Executor(t, r)
-	stage, err := builders.NewStageBuilder(r.Registry).
-		WithEncryptor(r.Encryptor).
-		InCanvas(r.Canvas.ID).
-		WithName("stage-1").
-		WithRequester(r.User).
-		WithConnections([]models.Connection{
-			{
-				SourceID:   r.Source.ID,
-				SourceType: models.SourceTypeEventSource,
-			},
-		}).
-		WithOutputs([]models.OutputDefinition{
-			{Name: "version", Required: true},
-			{Name: "sha", Required: true},
-		}).
-		WithExecutorType(executorType).
-		WithExecutorSpec(executorSpec).
-		ForResource(resource).
-		ForIntegration(r.Integration).
-		Create()
-
-	require.NoError(t, err)
-	signer := jwt.NewSigner("test")
-	server, err := NewServer(&crypto.NoOpEncryptor{}, r.Registry, signer, crypto.NewOIDCVerifier(), "", "", "/app/templates", authService)
-	require.NoError(t, err)
-
-	execution := support.CreateExecution(t, r.Source, stage)
-	superplaneToken, err := signer.Generate(execution.ID.String(), time.Hour)
-	require.NoError(t, err)
-
-	workflowID := uuid.NewString()
-	_, err = execution.AddResource(workflowID, "workflow", *stage.ResourceID)
-	require.NoError(t, err)
-
-	outputs := map[string]any{"version": "v1.0.0", "sha": "078fc8755c051"}
-
-	goodBody, _ := json.Marshal(&ExecutionOutputRequest{
-		ExternalID:  workflowID,
-		ExecutionID: execution.ID.String(),
-		Outputs:     outputs,
-	})
-
-	t.Run("event for invalid execution -> 404", func(t *testing.T) {
-		body, _ := json.Marshal(&ExecutionOutputRequest{
-			ExecutionID: "not-a-uuid",
-			Outputs:     outputs,
-		})
-
-		response := execRequest(server, requestParams{
-			method:      "POST",
-			path:        "/outputs",
-			body:        body,
-			authToken:   superplaneToken,
-			contentType: "application/json",
-		})
-
-		assert.Equal(t, 404, response.Code)
-		assert.Equal(t, "Execution not found\n", response.Body.String())
-	})
-
-	t.Run("missing Content-Type header -> 400", func(t *testing.T) {
-		response := execRequest(server, requestParams{
-			method:      "POST",
-			path:        "/outputs",
-			body:        goodBody,
-			contentType: "",
-			authToken:   superplaneToken,
-		})
-
-		assert.Equal(t, 404, response.Code)
-	})
-
-	t.Run("unsupported Content-Type header -> 400", func(t *testing.T) {
-		response := execRequest(server, requestParams{
-			method:      "POST",
-			path:        "/outputs",
-			body:        goodBody,
-			contentType: "application/x-www-form-urlencoded",
-			authToken:   superplaneToken,
-		})
-
-		assert.Equal(t, 404, response.Code)
-	})
-
-	t.Run("execution that does not exist -> 401", func(t *testing.T) {
-		body, _ := json.Marshal(&ExecutionOutputRequest{
-			ExecutionID: uuid.NewString(),
-			Outputs:     outputs,
-		})
-
-		response := execRequest(server, requestParams{
-			method:      "POST",
-			path:        "/outputs",
-			body:        body,
-			contentType: "application/json",
-			authToken:   superplaneToken,
-		})
-
-		assert.Equal(t, 404, response.Code)
-	})
-
-	t.Run("event with missing authorization header -> 401", func(t *testing.T) {
-		response := execRequest(server, requestParams{
-			method:      "POST",
-			path:        "/outputs",
-			body:        goodBody,
-			signature:   "",
-			authToken:   "",
-			contentType: "application/json",
-		})
-
-		assert.Equal(t, 401, response.Code)
-	})
-
-	t.Run("invalid auth token -> 403", func(t *testing.T) {
-		response := execRequest(server, requestParams{
-			method:      "POST",
-			path:        "/outputs",
-			body:        goodBody,
-			authToken:   "invalid",
-			contentType: "application/json",
-		})
-
-		assert.Equal(t, 401, response.Code)
-		assert.Equal(t, "Unauthorized\n", response.Body.String())
-	})
-
-	t.Run("superplane token is used -> 200 and execution outputs are updated", func(t *testing.T) {
-		response := execRequest(server, requestParams{
-			method:      "POST",
-			path:        "/outputs",
-			body:        goodBody,
-			authToken:   superplaneToken,
-			contentType: "application/json",
-		})
-
-		assert.Equal(t, 200, response.Code)
-		execution, err := models.FindExecutionByID(execution.ID, execution.StageID)
-		require.NoError(t, err)
-		assert.Equal(t, outputs, execution.Outputs.Data())
-	})
-
-	t.Run("integration OIDC ID token is used -> 200 and execution outputs are updated", func(t *testing.T) {
-		token := r.SemaphoreAPIMock.GenerateIDToken(resource.Id(), workflowID)
-		response := execRequest(server, requestParams{
-			method:      "POST",
-			path:        "/outputs",
-			body:        goodBody,
-			authToken:   token,
-			contentType: "application/json",
-		})
-
-		assert.Equal(t, 200, response.Code)
-		execution, err := models.FindExecutionByID(execution.ID, execution.StageID)
-		require.NoError(t, err)
-		assert.Equal(t, outputs, execution.Outputs.Data())
-	})
-
-	t.Run("output not defined in stage is ignored", func(t *testing.T) {
-		// 'time' output is not defined in the stage
-		body, _ := json.Marshal(&ExecutionOutputRequest{
-			ExecutionID: execution.ID.String(),
-			Outputs: map[string]any{
-				"sha":     "078fc8755c051",
-				"time":    1748555264,
-				"version": "v1.0.0",
-			},
-		})
-
-		response := execRequest(server, requestParams{
-			method:      "POST",
-			path:        "/outputs",
-			body:        body,
-			authToken:   superplaneToken,
-			contentType: "application/json",
-		})
-
-		assert.Equal(t, 200, response.Code)
-		execution, err := models.FindExecutionByID(execution.ID, execution.StageID)
-		require.NoError(t, err)
-		assert.Equal(t, map[string]any{"sha": "078fc8755c051", "version": "v1.0.0"}, execution.Outputs.Data())
-	})
-
-	t.Run("outputs are limited to 4k", func(t *testing.T) {
-		response := execRequest(server, requestParams{
-			method:      "POST",
-			path:        "/outputs",
-			body:        generateBigBody(t),
-			authToken:   superplaneToken,
-			contentType: "application/json",
-		})
-
-		assert.Equal(t, http.StatusRequestEntityTooLarge, response.Code)
-		assert.Equal(t, "Request body is too large - must be up to 4096 bytes\n", response.Body.String())
-	})
 }
 
 func Test__OpenAPIEndpoints(t *testing.T) {
