--- pkg/workers/hard_deletion_worker.go
@@ -1,271 +1,0 @@
-package workers
-
-import (
-	"errors"
-	"fmt"
-	"time"
-
-	log "github.com/sirupsen/logrus"
-	"github.com/superplanehq/superplane/pkg/database"
-	"github.com/superplanehq/superplane/pkg/models"
-	"github.com/superplanehq/superplane/pkg/registry"
-	"gorm.io/gorm"
-)
-
-const (
-	// Maximum number of items to process per tick
-	MaxHardDeletionBatchSize = 10
-)
-
-type HardDeletionWorker struct {
-	Registry       *registry.Registry
-	CleanupService *ResourceCleanupService
-	BatchSize      int
-}
-
-func NewHardDeletionWorker(registry *registry.Registry, cleanupService *ResourceCleanupService) *HardDeletionWorker {
-	return &HardDeletionWorker{
-		Registry:       registry,
-		CleanupService: cleanupService,
-		BatchSize:      MaxHardDeletionBatchSize,
-	}
-}
-
-func (w *HardDeletionWorker) Start() {
-	for {
-		err := w.Tick()
-		if err != nil {
-			log.Errorf("Error processing hard deletions: %v", err)
-		}
-
-		time.Sleep(30 * time.Second)
-	}
-}
-
-func (w *HardDeletionWorker) Tick() error {
-	// Process stages first since they have the most complex dependency chain
-	if err := w.processStages(); err != nil {
-		log.Errorf("Error processing stages for hard deletion: %v", err)
-	}
-
-	// Process event sources
-	if err := w.processEventSources(); err != nil {
-		log.Errorf("Error processing event sources for hard deletion: %v", err)
-	}
-
-	// Process connection groups
-	if err := w.processConnectionGroups(); err != nil {
-		log.Errorf("Error processing connection groups for hard deletion: %v", err)
-	}
-
-	return nil
-}
-
-// processStages handles hard deletion of soft-deleted stages following hierarchical dependency chain:
-// stages -> stage_events -> events, stage_executions -> execution_resources, connections
-func (w *HardDeletionWorker) processStages() error {
-	stages, err := models.ListUnscopedSoftDeletedStages(w.BatchSize)
-	if err != nil {
-		return fmt.Errorf("failed to list soft deleted stages: %v", err)
-	}
-
-	for _, stage := range stages {
-		logger := log.WithFields(log.Fields{
-			"stage_id":   stage.ID,
-			"stage_name": stage.Name,
-			"canvas_id":  stage.CanvasID,
-		})
-
-		logger.Info("Starting hard deletion of stage")
-
-		if err := w.hardDeleteStage(logger, &stage); err != nil {
-			logger.Errorf("Failed to hard delete stage: %v", err)
-		}
-	}
-
-	return nil
-}
-
-// hardDeleteStage handles hard deletion of soft-deleted stages following hierarchical dependency chain:
-// 1. Delete stage executions and their execution resources FIRST
-// (executions have foreign key to stage_events, so must be deleted first)
-// 2. Now delete stage events
-// 3. Delete connections where this stage is the target
-// 4. Clean up integration webhooks if stage has a resource
-// 5. Finally, hard delete the stage itself
-func (w *HardDeletionWorker) hardDeleteStage(logger *log.Entry, stage *models.Stage) error {
-	return database.Conn().Transaction(func(tx *gorm.DB) error {
-		if err := stage.DeleteStageExecutionsInTransaction(tx); err != nil {
-			return err
-		}
-
-		if err := stage.DeleteStageEventsInTransaction(tx); err != nil {
-			return err
-		}
-
-		if err := stage.DeleteConnectionsInTransaction(tx); err != nil {
-			return err
-		}
-
-		if stage.ResourceID != nil {
-			if err := w.CleanupService.CleanupStageWebhooksInTransaction(tx, stage); err != nil {
-				logger.Warnf("Failed to cleanup stage webhooks: %v", err)
-				// Don't fail the transaction, just log the warning
-			}
-		}
-
-		if err := stage.HardDeleteInTransaction(tx); err != nil {
-			return fmt.Errorf("failed to hard delete stage: %v", err)
-		}
-
-		logger.Info("Hard deleted stage with all dependencies")
-		return nil
-	})
-}
-
-// processEventSources handles hard deletion of soft-deleted event sources:
-// event_sources -> stage_events -> events, connections
-func (w *HardDeletionWorker) processEventSources() error {
-	eventSources, err := models.ListUnscopedSoftDeletedEventSources(w.BatchSize)
-	if err != nil {
-		return fmt.Errorf("failed to list soft deleted event sources: %v", err)
-	}
-
-	for _, eventSource := range eventSources {
-		logger := log.WithFields(log.Fields{
-			"event_source_id":   eventSource.ID,
-			"event_source_name": eventSource.Name,
-			"canvas_id":         eventSource.CanvasID,
-		})
-
-		logger.Info("Starting hard deletion of event source")
-
-		if err := w.hardDeleteEventSource(logger, &eventSource); err != nil {
-			logger.Errorf("Failed to hard delete event source: %v", err)
-		}
-	}
-
-	return nil
-}
-
-// hardDeleteEventSource handles hard deletion of soft-deleted event sources:
-// 1. Delete connections where this event source is the source
-// 2. Check if resource is still being used by other stages/event sources
-// 3. If resource is shared, convert event source to internal instead of hard deleting
-// 4. If resource is not shared, clean up webhooks and hard delete the event source
-func (w *HardDeletionWorker) hardDeleteEventSource(logger *log.Entry, eventSource *models.EventSource) error {
-	return database.Conn().Transaction(func(tx *gorm.DB) error {
-		if err := eventSource.DeleteConnectionsInTransaction(tx); err != nil {
-			return err
-		}
-
-		shouldDeleteResource, resource := w.prepareResourceCleanupInTransaction(tx, logger, eventSource)
-
-		//
-		// If resource is shared with other stages/event sources, convert to internal instead of deleting.
-		// So the webhooks are not deleted and keeps functional.
-		//
-		if resource != nil && !shouldDeleteResource {
-			logger.Info("Resource is shared, converting event source to internal instead of hard deleting")
-			if err := eventSource.ConvertToInternalInTransaction(tx); err != nil {
-				return fmt.Errorf("failed to convert event source to internal: %v", err)
-			}
-			logger.Info("Successfully converted event source to internal")
-			return nil
-		}
-
-		if err := eventSource.HardDeleteInTransaction(tx); err != nil {
-			return fmt.Errorf("failed to hard delete event source: %v", err)
-		}
-
-		//
-		// Now that the event source is deleted, we can safely delete the resource if it's safe to do so
-		//
-		if resource != nil && shouldDeleteResource {
-			if err := w.CleanupService.DeleteUnusedResourceWithChildrenInTransaction(tx, resource); err != nil {
-				logger.Errorf("Failed to cleanup unused resource: %v", err)
-				// Don't fail the transaction, just log the warning
-			}
-		}
-
-		logger.Info("Hard deleted event source with all dependencies")
-		return nil
-	})
-}
-
-func (w *HardDeletionWorker) prepareResourceCleanup(logger *log.Entry, eventSource *models.EventSource) (bool, *models.Resource) {
-	return w.prepareResourceCleanupInTransaction(database.Conn(), logger, eventSource)
-}
-
-func (w *HardDeletionWorker) prepareResourceCleanupInTransaction(tx *gorm.DB, logger *log.Entry, eventSource *models.EventSource) (bool, *models.Resource) {
-	if eventSource.ResourceID == nil {
-		return false, nil
-	}
-
-	resource, err := models.FindResourceByIDInTransaction(tx, *eventSource.ResourceID)
-	if err != nil {
-		if !errors.Is(err, gorm.ErrRecordNotFound) {
-			logger.Warnf("Failed to find resource for cleanup: %v", err)
-		}
-		return false, nil
-	}
-
-	canDelete, err := w.CleanupService.CleanupEventSourceWebhooksInTransaction(tx, eventSource, resource)
-	if err != nil {
-		logger.Warnf("Failed to cleanup event source webhooks: %v", err)
-		return false, resource
-	}
-
-	return canDelete, resource
-}
-
-// processConnectionGroups handles hard deletion of soft-deleted connection groups:
-// connection_groups -> connection_group_field_sets -> connection_group_field_set_events, connections
-func (w *HardDeletionWorker) processConnectionGroups() error {
-	connectionGroups, err := models.ListUnscopedSoftDeletedConnectionGroups(w.BatchSize)
-	if err != nil {
-		return fmt.Errorf("failed to list soft deleted connection groups: %v", err)
-	}
-
-	for _, connectionGroup := range connectionGroups {
-		logger := log.WithFields(log.Fields{
-			"connection_group_id":   connectionGroup.ID,
-			"connection_group_name": connectionGroup.Name,
-			"canvas_id":             connectionGroup.CanvasID,
-		})
-
-		logger.Info("Starting hard deletion of connection group")
-
-		if err := w.hardDeleteConnectionGroup(logger, &connectionGroup); err != nil {
-			logger.Errorf("Failed to hard delete connection group: %v", err)
-			continue
-		}
-
-		logger.Info("Successfully hard deleted connection group")
-	}
-
-	return nil
-}
-
-// hardDeleteConnectionGroup handles hard deletion of soft-deleted connection groups:
-// 1. Delete connection group field sets (but preserve their events)
-// 2. Delete connections where this connection group is the source or target
-// 3. Finally, hard delete the connection group itself
-func (w *HardDeletionWorker) hardDeleteConnectionGroup(logger *log.Entry, connectionGroup *models.ConnectionGroup) error {
-	return database.Conn().Transaction(func(tx *gorm.DB) error {
-		if err := connectionGroup.DeleteFieldSetsInTransaction(tx); err != nil {
-			return err
-		}
-
-		if err := connectionGroup.DeleteConnectionsInTransaction(tx); err != nil {
-			return err
-		}
-
-		if err := connectionGroup.HardDeleteInTransaction(tx); err != nil {
-			return fmt.Errorf("failed to hard delete connection group: %v", err)
-		}
-
-		logger.Info("Hard deleted connection group with all dependencies")
-		return nil
-	})
-}
