--- pkg/workers/resource_cleanup.go
@@ -1,195 +1,0 @@
-package workers
-
-import (
-	"context"
-	"errors"
-
-	"github.com/google/uuid"
-	log "github.com/sirupsen/logrus"
-	"github.com/superplanehq/superplane/pkg/database"
-	"github.com/superplanehq/superplane/pkg/models"
-	"github.com/superplanehq/superplane/pkg/registry"
-	"gorm.io/gorm"
-)
-
-type ResourceCleanupService struct {
-	Registry *registry.Registry
-}
-
-func NewResourceCleanupService(registry *registry.Registry) *ResourceCleanupService {
-	return &ResourceCleanupService{
-		Registry: registry,
-	}
-}
-
-func (s *ResourceCleanupService) CleanupUnusedResource(oldResourceID, excludeStageID uuid.UUID) error {
-	return s.CleanupUnusedResourceInTransaction(database.Conn(), oldResourceID, excludeStageID)
-}
-
-func (s *ResourceCleanupService) CleanupUnusedResourceInTransaction(tx *gorm.DB, oldResourceID, excludeStageID uuid.UUID) error {
-	oldResource, err := models.FindResourceByIDInTransaction(tx, oldResourceID)
-	if err != nil {
-		if errors.Is(err, gorm.ErrRecordNotFound) {
-			log.Warningf("Old resource %s not found, skipping cleanup", oldResourceID)
-			return nil
-		}
-		log.Errorf("Error finding old resource %s: %v", oldResourceID, err)
-		return err
-	}
-
-	logger := log.WithField("old_resource_id", oldResourceID)
-
-	externalEventSourceCount, err := models.CountExternalEventSourcesUsingResourceInTransaction(tx, oldResourceID)
-	if err != nil {
-		logger.Errorf("Error counting external event sources using resource: %v", err)
-		return err
-	}
-
-	stagesCount, err := models.CountOtherStagesUsingResourceInTransaction(tx, oldResourceID, excludeStageID)
-	if err != nil {
-		logger.Errorf("Error counting stages using resource: %v", err)
-		return err
-	}
-
-	totalUsages := externalEventSourceCount + stagesCount
-	if totalUsages > 0 {
-		logger.Infof("Resource is used by %d external event sources and %d stages, skipping cleanup",
-			externalEventSourceCount, stagesCount)
-		return nil
-	}
-
-	integration, err := models.FindIntegrationByIDInTransaction(tx, oldResource.IntegrationID)
-	if err != nil {
-		if errors.Is(err, gorm.ErrRecordNotFound) {
-			logger.Info("Integration not found, skipping cleanup")
-			return nil
-		}
-		logger.Errorf("Error finding integration: %v", err)
-		return err
-	}
-
-	resourceManager, err := s.Registry.NewResourceManagerInTransaction(context.Background(), tx, integration)
-	if err != nil {
-		logger.Errorf("Error creating resource manager: %v", err)
-		return err
-	}
-
-	childResources, err := oldResource.FindChildrenInTransaction(tx)
-	if err != nil {
-		logger.Errorf("Error finding child resources: %v", err)
-		return err
-	}
-
-	for _, childResource := range childResources {
-		logger.Infof("Cleaning up webhook resource %s", childResource.Id())
-		err = resourceManager.CleanupWebhook(oldResource, &childResource)
-		if err != nil {
-			logger.Errorf("Error cleaning up webhook: %v", err)
-		} else {
-			logger.Infof("Successfully cleaned up webhook %s", childResource.Id())
-		}
-	}
-
-	err = models.DeleteResourceWithChildrenInTransaction(tx, oldResourceID)
-	if err != nil {
-		logger.Errorf("Error deleting old resource and children from database: %v", err)
-		return err
-	}
-
-	logger.Infof("Successfully cleaned up and deleted old resource %s and %d child resources", oldResourceID, len(childResources))
-
-	return nil
-}
-
-func (s *ResourceCleanupService) DeleteUnusedResourceWithChildren(oldResource *models.Resource) error {
-	return s.DeleteUnusedResourceWithChildrenInTransaction(database.Conn(), oldResource)
-}
-
-func (s *ResourceCleanupService) DeleteUnusedResourceWithChildrenInTransaction(tx *gorm.DB, oldResource *models.Resource) error {
-	logger := log.WithField("old_resource_id", oldResource.ID)
-
-	err := models.DeleteResourceWithChildrenInTransaction(tx, oldResource.ID)
-	if err != nil {
-		logger.Errorf("Error deleting old resource and children from database: %v", err)
-		return err
-	}
-
-	logger.Infof("Successfully deleted old resource %s from database", oldResource.ID)
-
-	return nil
-}
-
-func (s *ResourceCleanupService) CleanupStageWebhooksInTransaction(tx *gorm.DB, stage *models.Stage) error {
-	if stage.ResourceID == nil {
-		return nil
-	}
-
-	// Use the existing resource cleanup service to handle webhook cleanup
-	// We pass uuid.Nil as excludeStageID since we're deleting the stage itself
-	return s.CleanupUnusedResourceInTransaction(tx, *stage.ResourceID, uuid.Nil)
-}
-
-func (s *ResourceCleanupService) CleanupEventSourceWebhooks(eventSource *models.EventSource, resource *models.Resource) (bool, error) {
-	return s.CleanupEventSourceWebhooksInTransaction(database.Conn(), eventSource, resource)
-}
-
-func (s *ResourceCleanupService) CleanupEventSourceWebhooksInTransaction(tx *gorm.DB, eventSource *models.EventSource, resource *models.Resource) (bool, error) {
-	if eventSource.ResourceID == nil {
-		return false, nil
-	}
-
-	if resource == nil {
-		return false, nil
-	}
-
-	logger := log.WithField("event_source_id", eventSource.ID)
-
-	externalEventSourceCount, err := models.CountExternalEventSourcesUsingResourceInTransaction(tx, *eventSource.ResourceID)
-	if err != nil {
-		return false, err
-	}
-
-	stagesCount, err := models.CountStagesUsingResourceInTransaction(tx, *eventSource.ResourceID)
-	if err != nil {
-		return false, err
-	}
-
-	logger.Infof("Resource %s is used by %d external event sources and %d stages", *eventSource.ResourceID, externalEventSourceCount, stagesCount)
-
-	totalUsages := externalEventSourceCount + stagesCount
-
-	//
-	// Soft deleted event sources are not considered in scoped queries,
-	// So we dont consider them in the total usages count
-	//
-	if totalUsages > 0 {
-		return false, nil
-	}
-
-	integration, err := models.FindIntegrationByIDInTransaction(tx, resource.IntegrationID)
-	if err != nil {
-		return false, err
-	}
-
-	resourceManager, err := s.Registry.NewResourceManagerInTransaction(context.Background(), tx, integration)
-	if err != nil {
-		return false, err
-	}
-
-	childResources, err := resource.FindChildrenInTransaction(tx)
-	if err != nil {
-		return false, err
-	}
-
-	for _, childResource := range childResources {
-		logger.Infof("Cleaning up webhook resource %s", childResource.Id())
-		err = resourceManager.CleanupWebhook(resource, &childResource)
-		if err != nil {
-			logger.Errorf("Error cleaning up webhook: %v", err)
-		} else {
-			logger.Infof("Successfully cleaned up webhook %s", childResource.Id())
-		}
-	}
-
-	return true, nil
-}
